added back in one hack to get debugging working again
One hack still needed - the 'sources' field in the bundled sourcemap by
default has a leading slash; but chrome fails to find the sourcemap in
that case.  I can't find the  right combination of src.base, file
paths, etc to get it to work, so just removing the slash manually via
mapSources.