Dependent projects can now have their js files copied automatically with copyDependencyLibs
project.copyDependencyLibs:boolean ==  Project A can define 'dependsOn'
projects [B,C,...] as per above; those dependent projects' d.ts files
are automatically copied over to project A so that it gets ambient
typing info.  By default, we don't copy the dependent projects' output
js files over, as we assume that project A will reference it in /dist
directly.  This
approach minimizes duplicate copies (and git messiness) for a project
with libs and samples.  However, if you have a setup like that, and the
samples are intended to be completely standalone (e.g. a webserver
points directly at their folder, and they can't see "/dist"), then you
can override that by setting copyDependencyLibs to true.  You can do
this directly on an individual project, or in the project
defaults here.   You'll get duplicate copies of 'projectB-debug.js' in
your repo, but project A will be completely standalone.  Defaults to
false